Structuring HD , Fs ...

+++ enlargeArray.h
@@ -5,8 +5,8 @@
 #define NONE *****
 #define offset 0.5
 
-void reallocSarray(IndexElemC* sarray, long* sarray_len);
+void reallocSarray(IndexElemC* sarray, int *sarray_len);
 
-void reallocFileIndex(IndexElemC* sarray, long* sarray_len, int _id);
+void reallocFileIndex(IndexElemC* sarray, int* sarray_len, int _id);
 
 #endif // ENLARGEARRAY_H
